pulling logic out of Container into specific components
