+++ src/client/screens/Workout/Walk/WalkWorkout.tsx
@@ -6,10 +6,10 @@ type WalkProps = {
 
 const WalkWorkout = ({ workout }: WalkProps) => {
   return (
-    <>
+    <div className="col-span-12">
       <div className="text-4xl">Walk</div>
       {JSON.stringify(workout)}
-    </>
+    </div>
   );
 };
